Refactor ClassLoaderUtil into its own exported package.
