+++ quasar-core/src/main/java/co/paralleluniverse/common/asm/ASMUtil.java
@@ -13,8 +13,8 @@ import java.lang.reflect.Member;
 import java.lang.reflect.Method;
 import java.util.concurrent.atomic.AtomicReference;
 
-import static co.paralleluniverse.asm.ASMUtil.classToSlashed;
 import static co.paralleluniverse.asm.ASMUtil.getClassNode;
+import static co.paralleluniverse.common.resource.ClassLoaderUtil.classToSlashed;
 import static org.objectweb.asm.ClassReader.SKIP_FRAMES;
 
 public final class ASMUtil {
