+++ quasar-core/src/main/java/co/paralleluniverse/common/resource/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package co.paralleluniverse.common.resource;
+
+import org.osgi.annotation.bundle.Export;
