--- quasar-core/src/main/java/co/paralleluniverse/fibers/instrument/ClassLoaderUtil.java
@@ -1,60 +1,0 @@
-/*
- * Quasar: lightweight threads and actors for the JVM.
- * Copyright (c) 2013-2018, Parallel Universe Software Co. All rights reserved.
- *
- * This program and the accompanying materials are dual-licensed under
- * either the terms of the Eclipse Public License v1.0 as published by
- * the Eclipse Foundation
- *
- *   or (per the licensee's choosing)
- *
- * under the terms of the GNU Lesser General Public License version 3.0
- * as published by the Free Software Foundation.
- */
-/*
- * Based on Guava's com.google.common.reflect.ClassPath
- */
-/*
- * Copyright (c) 2012 The Guava Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package co.paralleluniverse.fibers.instrument;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.URL;
-import java.net.URLConnection;
-import java.util.Enumeration;
-
-final class ClassLoaderUtil {
-    public static URL getResource(ClassLoader cl, String resource) {
-        return cl != null ? cl.getResource(resource) : ClassLoader.getSystemResource(resource);
-    }
-
-    public static Enumeration<URL> getResources(ClassLoader cl, String resources) throws IOException {
-        return cl != null ? cl.getResources(resources) : ClassLoader.getSystemResources(resources);
-    }
-
-    public static InputStream getResourceAsStream(ClassLoader cl, String resource) throws IOException {
-        URL url = getResource(cl, resource);
-        if (url == null)
-            return null;
-        URLConnection uc = url.openConnection();
-        uc.setUseCaches(false);
-        return uc.getInputStream();
-    }
-
-    private ClassLoaderUtil() {
-    }
-}
