+++ quasar-core/src/main/java/co/paralleluniverse/fibers/instrument/QuasarInstrumentor.java
@@ -40,7 +40,8 @@ public final class QuasarInstrumentor {
 
     private static final List<String> BUILT_IN_PACKAGES = List.of(
         "co/paralleluniverse/asm/",
-        "co/paraleluniverse/common/asm/",
+        "co/paralleluniverse/common/asm/",
+        "co/paralleluniverse/common/resource/",
         "org/objectweb/asm/", // For testing
         "org/netbeans/lib/"
     );
